Refactored actions of deployment processing
