+++ src/main/java/org/qubership/integration/platform/engine/service/deployment/processing/DeploymentProcessingAction.java
@@ -1,0 +1,13 @@
+package org.qubership.integration.platform.engine.service.deployment.processing;
+
+import org.apache.camel.spring.SpringCamelContext;
+import org.qubership.integration.platform.engine.model.deployment.update.DeploymentConfiguration;
+import org.qubership.integration.platform.engine.model.deployment.update.DeploymentInfo;
+
+public interface DeploymentProcessingAction {
+    void execute(
+        SpringCamelContext context,
+        DeploymentInfo deploymentInfo,
+        DeploymentConfiguration deploymentConfiguration
+    );
+}
