+++ src/main/java/org/qubership/integration/platform/engine/service/deployment/processing/actions/context/create/helpers/MetricTagsHelper.java
@@ -1,0 +1,60 @@
+package org.qubership.integration.platform.engine.service.deployment.processing.actions.context.create.helpers;
+
+import org.qubership.integration.platform.engine.model.constants.CamelConstants.ChainProperties;
+import org.qubership.integration.platform.engine.model.deployment.update.DeploymentInfo;
+import org.qubership.integration.platform.engine.model.deployment.update.ElementProperties;
+import org.qubership.integration.platform.engine.configuration.ServerConfiguration;
+import io.micrometer.core.instrument.Tag;
+import io.micrometer.common.KeyValue;
+import io.micrometer.common.KeyValues;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collection;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+
+import static org.qubership.integration.platform.engine.service.debugger.metrics.MetricsStore.CHAIN_ID_TAG;
+import static org.qubership.integration.platform.engine.service.debugger.metrics.MetricsStore.CHAIN_NAME_TAG;
+import static org.qubership.integration.platform.engine.service.debugger.metrics.MetricsStore.ELEMENT_ID_TAG;
+import static org.qubership.integration.platform.engine.service.debugger.metrics.MetricsStore.ELEMENT_NAME_TAG;
+import static org.qubership.integration.platform.engine.service.debugger.metrics.MetricsStore.ENGINE_DOMAIN_TAG;
+
+@Component
+public class MetricTagsHelper {
+    public final ServerConfiguration serverConfiguration;
+
+    @Autowired
+    public MetricTagsHelper(ServerConfiguration serverConfiguration) {
+        this.serverConfiguration = serverConfiguration;
+    }
+
+    public Collection<Tag> buildMetricTagsLegacy(
+        DeploymentInfo deploymentInfo,
+        ElementProperties elementProperties,
+        String chainName
+    ) {
+        return new ArrayList<>(Arrays.asList(
+            Tag.of(CHAIN_ID_TAG, deploymentInfo.getChainId()),
+            Tag.of(CHAIN_NAME_TAG, chainName),
+            Tag.of(ELEMENT_ID_TAG, elementProperties.getProperties().get(ChainProperties.ELEMENT_ID)),
+            Tag.of(ELEMENT_NAME_TAG, elementProperties.getProperties().get(
+                ChainProperties.ELEMENT_NAME)),
+            Tag.of(ENGINE_DOMAIN_TAG, serverConfiguration.getDomain()))
+        );
+    }
+
+    public KeyValues buildMetricTags(
+        DeploymentInfo deploymentInfo,
+        ElementProperties elementProperties,
+        String chainName
+    ) {
+        return KeyValues.of(
+            KeyValue.of(CHAIN_ID_TAG, deploymentInfo.getChainId()),
+            KeyValue.of(CHAIN_NAME_TAG, chainName),
+            KeyValue.of(ELEMENT_ID_TAG, elementProperties.getProperties().get(
+                ChainProperties.ELEMENT_ID)),
+            KeyValue.of(ELEMENT_NAME_TAG, elementProperties.getProperties().get(
+                ChainProperties.ELEMENT_NAME)),
+            KeyValue.of(ENGINE_DOMAIN_TAG, serverConfiguration.getDomain()));
+    }
+}
